Refactor email methods into their own class
