+++ app/workers/end_of_cycle/send_reject_by_default_reminder_to_providers_worker.rb
@@ -22,7 +22,7 @@ module EndOfCycle
   private
 
     def do_not_send_email
-      !CycleTimetable.send_reject_by_default_reminder_to_providers?
+      !EmailTimetable.send_reject_by_default_reminder_to_providers?
     end
   end
